pyproject.toml
[tool.poetry]
name = "fal-ext"
version = "0.0.1"
description = "`fal-ext` is a Meltano utility extension."
authors = ["Stephane Burwash"]
license = "Apache 2.0"
packages = [
    { include = "fal_ext" }
    # uncomment this if you want to include static assets
    # { include = "files_fal-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_fal-ext/"
#include = [
#    "files_fal-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = ">=3.7.2,<3.11"
structlog = "^21.2.0"
PyYAML = "^6.0.0"
pydantic = "^1.9.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="main"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fal_extension = 'fal_ext.main:app'
fal_invoker = 'fal_ext.pass_through:pass_through_cli'
